requirements.txt
scikit-learn==1.4.2
mlflow==2.13.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
evidently==0.4.30
dvc==3.50.1
pandas==2.2.2
numpy==1.26.4
pytest==8.2.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.1
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
